moving logics under modules - all extract_faces logic moved to detection module - load image related logics moved to preprocessing module
