+++ deepface/modules/detection.py
@@ -3,10 +3,26 @@ from typing import Any, Dict, List, Tuple, Union
 
 # 3rd part dependencies
 import numpy as np
+import cv2
 from PIL import Image
 
 # project dependencies
+from deepface.modules import preprocessing
+from deepface.models.Detector import DetectedFace, FacialAreaRegion
+from deepface.detectors import DetectorWrapper
 from deepface.commons import functions
+from deepface.commons.logger import Logger
+
+logger = Logger(module="deepface/modules/detection.py")
+
+# pylint: disable=no-else-raise
+
+
+tf_major_version = functions.get_tf_major_version()
+if tf_major_version == 1:
+    from keras.preprocessing import image
+elif tf_major_version == 2:
+    from tensorflow.keras.preprocessing import image
 
 
 def extract_faces(
@@ -16,6 +32,7 @@ def extract_faces(
     enforce_detection: bool = True,
     align: bool = True,
     grayscale: bool = False,
+    human_readable=False,
 ) -> List[Dict[str, Any]]:
     """
     Extract faces from a given image
@@ -38,6 +55,8 @@ def extract_faces(
         grayscale (boolean): Flag to convert the image to grayscale before
             processing (default is False).
 
+        human_readable (bool): Flag to make the image human readable. 3D RGB for human readable
+            or 4D BGR for ML models (default is False).
 
     Returns:
         results (List[Dict[str, Any]]): A list of dictionaries, where each dictionary contains:
@@ -48,27 +67,108 @@ def extract_faces(
 
     resp_objs = []
 
-    img_objs = functions.extract_faces(
-        img=img_path,
-        target_size=target_size,
-        detector_backend=detector_backend,
-        grayscale=grayscale,
-        enforce_detection=enforce_detection,
-        align=align,
-    )
+    # img might be path, base64 or numpy array. Convert it to numpy whatever it is.
+    img, img_name = preprocessing.load_image(img_path)
 
-    for img, region, confidence in img_objs:
-        resp_obj = {}
+    base_region = FacialAreaRegion(x=0, y=0, w=img.shape[1], h=img.shape[0])
 
+    if detector_backend == "skip":
+        face_objs = [DetectedFace(img=img, facial_area=base_region, confidence=0)]
+    else:
+        face_objs = DetectorWrapper.detect_faces(detector_backend, img, align)
+
+    # in case of no face found
+    if len(face_objs) == 0 and enforce_detection is True:
+        if img_name is not None:
+            raise ValueError(
+                f"Face could not be detected in {img_name}."
+                "Please confirm that the picture is a face photo "
+                "or consider to set enforce_detection param to False."
+            )
+        else:
+            raise ValueError(
+                "Face could not be detected. Please confirm that the picture is a face photo "
+                "or consider to set enforce_detection param to False."
+            )
+
+    if len(face_objs) == 0 and enforce_detection is False:
+        face_objs = [DetectedFace(img=img, facial_area=base_region, confidence=0)]
+
+    for face_obj in face_objs:
+        current_img = face_obj.img
+        current_region = face_obj.facial_area
+        confidence = face_obj.confidence
+
+        if current_img.shape[0] == 0 or current_img.shape[1] == 0:
+            continue
+
+        if grayscale is True:
+            current_img = cv2.cvtColor(current_img, cv2.COLOR_BGR2GRAY)
+
+        # resize and padding
+        factor_0 = target_size[0] / current_img.shape[0]
+        factor_1 = target_size[1] / current_img.shape[1]
+        factor = min(factor_0, factor_1)
+
+        dsize = (
+            int(current_img.shape[1] * factor),
+            int(current_img.shape[0] * factor),
+        )
+        current_img = cv2.resize(current_img, dsize)
+
+        diff_0 = target_size[0] - current_img.shape[0]
+        diff_1 = target_size[1] - current_img.shape[1]
+        if grayscale is False:
+            # Put the base image in the middle of the padded image
+            current_img = np.pad(
+                current_img,
+                (
+                    (diff_0 // 2, diff_0 - diff_0 // 2),
+                    (diff_1 // 2, diff_1 - diff_1 // 2),
+                    (0, 0),
+                ),
+                "constant",
+            )
+        else:
+            current_img = np.pad(
+                current_img,
+                (
+                    (diff_0 // 2, diff_0 - diff_0 // 2),
+                    (diff_1 // 2, diff_1 - diff_1 // 2),
+                ),
+                "constant",
+            )
+
+        # double check: if target image is not still the same size with target.
+        if current_img.shape[0:2] != target_size:
+            current_img = cv2.resize(current_img, target_size)
+
+        # normalizing the image pixels
+        # what this line doing? must?
+        img_pixels = image.img_to_array(current_img)
+        img_pixels = np.expand_dims(img_pixels, axis=0)
+        img_pixels /= 255  # normalize input in [0, 1]
         # discard expanded dimension
-        if len(img.shape) == 4:
-            img = img[0]
+        if human_readable is True and len(img_pixels.shape) == 4:
+            img_pixels = img_pixels[0]
 
-        # bgr to rgb
-        resp_obj["face"] = img[:, :, ::-1]
-        resp_obj["facial_area"] = region
-        resp_obj["confidence"] = confidence
-        resp_objs.append(resp_obj)
+        resp_objs.append(
+            {
+                "face": img_pixels[:, :, ::-1] if human_readable is True else img_pixels,
+                "facial_area": {
+                    "x": current_region.x,
+                    "y": current_region.y,
+                    "w": current_region.w,
+                    "h": current_region.h,
+                },
+                "confidence": confidence,
+            }
+        )
+
+    if len(resp_objs) == 0 and enforce_detection == True:
+        raise ValueError(
+            f"Detected face shape is {img.shape}. Consider to set enforce_detection arg to False."
+        )
 
     return resp_objs
